FIX : integer divided by integer give integer: so 0,x value id rounded to 0 making Text (Class=4, Element=4) to be invisible

+++ src/net/sf/jcgm/core/TextCommand.java
@@ -126,7 +126,7 @@ public abstract class TextCommand extends Command {
 			screenResolution = Toolkit.getDefaultToolkit()
 					.getScreenResolution();
 		}
-		double height = fontMetrics.getAscent() * 72 / screenResolution;
+		double height = fontMetrics.getAscent() * 72 / (double) screenResolution;
 
 		scaleText(d, fontMetrics, glyphVector, logicalBounds.getWidth(), height);
